requirements.txt
openai
python-dotenv
spacy

streamlit
python-docx
PyMuPDF
ollama